+++ x/erc20/types/codec.go
@@ -26,6 +26,7 @@ const (
 	// Amino names
 	convertERC20Name = "evmos/MsgConvertERC20"
 	convertCoinName  = "evmos/MsgConvertCoin"
+	updateParams     = "evmos/erc20/MsgUpdateParams"
 )
 
 // NOTE: This is required for the GetSignBytes function
@@ -40,6 +41,7 @@ func RegisterInterfaces(registry codectypes.InterfaceRegistry) {
 		(*sdk.Msg)(nil),
 		&MsgConvertCoin{},
 		&MsgConvertERC20{},
+		&MsgUpdateParams{},
 	)
 	registry.RegisterImplementations(
 		(*govv1beta1.Content)(nil),
@@ -55,6 +57,7 @@ func RegisterInterfaces(registry codectypes.InterfaceRegistry) {
 // concrete types on the provided LegacyAmino codec. These types are used for
 // Amino JSON serialization and EIP-712 compatibility.
 func RegisterLegacyAminoCodec(cdc *codec.LegacyAmino) {
+	cdc.RegisterConcrete(&MsgUpdateParams{}, updateParams, nil)
 	cdc.RegisterConcrete(&MsgConvertERC20{}, convertERC20Name, nil)
 	cdc.RegisterConcrete(&MsgConvertCoin{}, convertCoinName, nil)
 }